requirements.txt
Flask==1.1.1
pyserial==3.4
six==1.13.0
